analytics-pipeline/requirements.txt
google-ads==24.1.0
google-analytics-data==0.18.13
google-api-python-client==2.136.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.25.0
python-dotenv==1.0.1
pandas==2.2.2
pyarrow==17.0.0
